+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release 0.1.6
 
 * Properties could hava a type
+* Check if a method is not yet implemented
 
 # Release 0.1.5